+++ packages/fury-adapter-oas3-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Fury OAS3 Parser Changelog
 
+## Master
+
+### Enhancements
+
+- Object properties in a Schema Object are now supported.
+
 ## 0.4.1 (28-01-19)
 
 ### Enhancements